LayoutTests/platform/glib/imported/w3c/web-platform-tests/focus/focus-event-after-iframe-gets-focus-expected.txt


PASS Check focus event after same site iframe gets focus
FAIL Check focus event after different site iframe gets focus assert_equals: expected "outerlog:windowblur,innerlog:willfocuswindow,windowfocus,didfocuswindow," but got "outerlog:innerlog:willfocuswindow,didfocuswindow,"
FAIL Check focus event after innermost different site iframe gets focus assert_equals: expected "outerlog:windowblur,middlelog:innerlog:willfocuswindow,windowfocus,didfocuswindow," but got "outerlog:middlelog:innerlog:willfocuswindow,didfocuswindow,"
